+++ src/Raffle.sol
@@ -126,11 +126,11 @@ contract Raffle is VRFConsumerBaseV2 {
         return i_entranceFee;
     }
 
-     function getRaffleState() external view returns (RaffleState) {
+    function getRaffleState() external view returns (RaffleState) {
         return s_raffleState;
     }
 
-    function getPlayer(uint index) public view returns(address) {
+    function getPlayer(uint256 index) public view returns (address) {
         return s_players[index];
     }
 }
